+++ libvast/vast/type.hpp
@@ -30,14 +30,13 @@
 
 #include <caf/detail/type_list.hpp>
 #include <caf/error.hpp>
-#include <caf/expected.hpp>
 #include <caf/fwd.hpp>
 #include <caf/intrusive_cow_ptr.hpp>
 #include <caf/make_counted.hpp>
 #include <caf/meta/omittable.hpp>
 #include <caf/none.hpp>
 #include <caf/ref_counted.hpp>
-#include <caf/variant.hpp>
+#include <caf/sum_type.hpp>
 
 #include <functional>
 #include <string>
@@ -301,6 +300,32 @@ public:
 
   Derived attributes(std::vector<attribute> xs) && {
     this->attributes_ = std::move(xs);
+    return std::move(derived());
+  }
+
+  Derived& update_attributes(std::vector<attribute> xs) & {
+    auto& attrs = this->attributes_;
+    for (auto& x : xs) {
+      auto i = std::find_if(attrs.begin(), attrs.end(),
+                            [&](auto& attr) { return attr.key == x.key; });
+      if (i == attrs.end())
+        attrs.push_back(std::move(x));
+      else
+        i->value = std::move(x).value;
+    }
+    return derived();
+  }
+
+  Derived update_attributes(std::vector<attribute> xs) && {
+    auto& attrs = this->attributes_;
+    for (auto& x : xs) {
+      auto i = std::find_if(attrs.begin(), attrs.end(),
+                            [&](auto& attr) { return attr.key == x.key; });
+      if (i == attrs.end())
+        attrs.push_back(std::move(x));
+      else
+        i->value = std::move(x).value;
+    }
     return std::move(derived());
   }
 
@@ -718,8 +743,42 @@ record_type concat(const Rs&... rs) {
   (result.fields.insert(result.fields.end(), rs.fields.begin(),
                         rs.fields.end()),
    ...);
+  // TODO: This function is missing an integrity check that makes sure the
+  // result does not contain multiple fields with the same name.
+  // We should also add a differently named version that deduplicates completely
+  // identical fields and recurses into nested records under the same field
+  // name.
   return result;
 }
+
+/// Creates a new unnamed record_type containing the fields and attribues of lhs
+/// and rhs. Errors if a field of the same name but different types is present
+/// in both inputs. Errors is the inputs disagree over the value of an attribute
+/// with a certain name.
+/// @returns The combined record_type.
+/// @relates record_type
+caf::expected<record_type>
+merge(const record_type& lhs, const record_type& rhs);
+
+/// @relates priority_merge
+enum class merge_policy { prefer_left, prefer_right };
+
+/// Creates a new unnamed record_type containing the fields and attribues of lhs
+/// and rhs. Uses a merge_policy to decide wheter to use a field from lhs or rhs
+/// in case of a conflict.
+/// @returns The combined record_type.
+/// @relates record_type merge_policy
+record_type
+priority_merge(const record_type& lhs, const record_type& rhs, merge_policy p);
+
+/// Removes a field from a record_type by name.
+/// @param r The record to mutate.
+/// @param path The sequence of keys pointing to the target field.
+/// @returns A bool indicating whether a field of the name field_name was
+///          present before this function was called.
+/// @pre `!path.empty()`
+/// @relates record_type
+bool remove_field(record_type& r, std::vector<std::string_view> path);
 
 /// Recursively flattens the arguments of a record type.
 /// @param rec The record to flatten.
@@ -940,6 +999,14 @@ struct sum_type_access<vast::type> {
   template <class T, int Pos>
   static const T& get(const vast::type& x, sum_type_token<T, Pos>) {
     return static_cast<const T&>(*x);
+  }
+
+  template <class T, int Pos>
+  static T* get_if(vast::type* x, sum_type_token<T, Pos>) {
+    x->ptr().unshare();
+    auto ptr = x->raw_ptr();
+    return ptr->index() == Pos ? const_cast<T*>(static_cast<const T*>(ptr))
+                               : nullptr;
   }
 
   template <class T, int Pos>